refactor: adapt handlers to new services, rename files, and improve types
