+++ functions/core/media-upload.ts
@@ -1,6 +1,6 @@
 import { S3Client, PutObjectCommand } from '@aws-sdk/client-s3';
 import { getSignedUrl } from '@aws-sdk/s3-request-presigner';
-import { ApiResponse, success, error } from '../shared/responses';
+import { ApiResponse, error, success } from '../shared/dynamo';
 
 const s3 = new S3Client({ region: process.env.AWS_REGION });
