--- functions/therapies/create-therapy.ts
@@ -1,43 +1,0 @@
-import { PutCommand } from '@aws-sdk/lib-dynamodb';
-import { v4 as uuidv4 } from 'uuid';
-import { docClient } from "../handlers/shared/db-client";
-import { ApiResponse, error, success } from '../handlers/shared/responses';
-import { CreateTherapyInput } from '../handlers/shared/types/therapy';
-
-export const handler = async (event: { body?: string; }): Promise<ApiResponse> => {
-  const input = JSON.parse(event.body || '{}') as CreateTherapyInput;
-
-  if (!input.title || !input.description || !input.content) {
-    return error(400, 'Missing required fields');
-  }
-
-  try {
-    const therapyId = uuidv4();
-
-    await docClient.send(
-      new PutCommand({
-        TableName: process.env.TABLE_NAME,
-        Item: {
-          PK: `THERAPY#${therapyId}`,
-          SK: `THERAPY#${therapyId}`,
-          Type: 'Therapy',
-          therapyId,
-          title: input.title,
-          description: input.description,
-          content: input.content,
-          image: input.image,
-          isGroup: input.isGroup,
-          createdAt: new Date().toISOString(),
-        }
-      })
-    );
-
-    return success({
-      message: 'Therapy created successfully',
-      therapyId
-    });
-  } catch(err) {
-    console.error('Error creting therapy', err);
-    return error(500, 'Internal Server Error');
-  }
-};
